Move metric names back
